+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-03-17
+
+### Added
+
+- Added react-native-vector-icons as a dependency
+- Added @types/react-native-vector-icons as a dev dependency
+- Added icon support to the components
+- Added dependency detection to the add command (dependency.json logic) so that it can install other components that are dependencies of the component being added
+- Added dependency detection to the init command so that it can install other dependencies
+- Added accordion, alert, alertdialog
+
 ## [0.1.11] - 2025-03-16
 
 ### Changed